refactor: rework how symlinks are processed (no longer resolve) (#248)
Some changes to how we handle symlinks. We no longer resolve them, which should reduce the complexity by a nice bit.

We also now show the target name in the output.

I also added a launch.json file for debugging because it took me a while to figure out how to get the debugger to work.

Yeah, that's it.

Please test before merging because I'm a bit of a dingus sometimes